+++ src/app/dashboard/document/documentComponents/TabsDocuments.tsx
@@ -1,21 +1,7 @@
 'use client';
 import DocumentNav from '@/components/DocumentNav';
-import NewDocumentType from '@/components/NewDocumentType';
-import {
-  AlertDialog,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from '@/components/ui/alert-dialog';
-import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
-import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/ui/table';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { useCountriesStore } from '@/store/countries';
 import { CompanyDocumentsType, useLoggedUserStore } from '@/store/loggedUser';
 import cookies from 'js-cookie';
 import { useRouter } from 'next/navigation';
@@ -25,7 +11,7 @@ import { ColumnsMonthly } from '../../columsMonthly';
 import { DataTable } from '../../company/actualCompany/components/data-table';
 import { columnsDocuments } from '../../company/actualCompany/components/document-colums';
 import { ExpiredDataTable } from '../../data-table';
-import { EditModal } from './EditDocumenTypeModal';
+import TypesDocumentsView from './TypesDocumentsView';
 
 interface Document {
   date: string;
@@ -57,21 +43,15 @@ function TabsDocuments({
   };
   clientData: any[] | null;
 }) {
-  const { actualCompany } = useLoggedUserStore();
   const actualComp = cookies.get('actualComp');
 
   useEffect(() => {
     router.refresh();
   }, [actualComp]);
 
-  const document_types = useCountriesStore((state) => state.companyDocumentTypes);
-  let doc_personas = document_types?.filter((doc) => doc.applies === 'Persona').filter((e) => e.is_active);
-  let doc_equipos = document_types?.filter((doc) => doc.applies === 'Equipos').filter((e) => e.is_active);
-  let doc_empresa = document_types?.filter((doc) => doc.applies === 'Empresa').filter((e) => e.is_active);
   const profile = useLoggedUserStore((state) => state);
   const sharedUsersAll = useLoggedUserStore((state) => state.sharedUsers);
   const role = serverRole ?? useLoggedUserStore((state) => state.roleActualCompany);
-  const fetchDocumentTypes = useCountriesStore((state) => state.documentTypes);
   const ownerUser = useLoggedUserStore((state) => state.profile);
   const sharedUsers =
     sharedUsersAll?.map((user) => {
@@ -358,194 +338,7 @@ function TabsDocuments({
         </Card>
       </TabsContent>
       <TabsContent value="Tipos de documentos">
-        <Card>
-          <div className="flex justify-between items-center">
-            <CardHeader>
-              <CardTitle className="text-2xl font-bold tracking-tight">Tipos de documentos</CardTitle>
-              <CardDescription className="text-muted-foreground">Tipos de documentos auditables</CardDescription>
-            </CardHeader>
-
-            <AlertDialog>
-              <AlertDialogTrigger asChild className="mr-4">
-                {role !== 'Invitado' && <Button>Crear nuevo</Button>}
-              </AlertDialogTrigger>
-              <AlertDialogContent>
-                <AlertDialogHeader>
-                  <AlertDialogTitle>Nuevo tipo de documento</AlertDialogTitle>
-                  <AlertDialogDescription asChild>
-                    <NewDocumentType codeControlClient />
-                  </AlertDialogDescription>
-                </AlertDialogHeader>
-                <AlertDialogFooter>
-                  <Button
-                    onClick={() => {
-                      document.getElementById('create_new_document')?.click();
-                      fetchDocumentTypes(actualCompany?.id);
-                    }}
-                  >
-                    Crear documento
-                  </Button>
-                  <AlertDialogCancel id="close_document_modal">Cancel</AlertDialogCancel>
-                </AlertDialogFooter>
-              </AlertDialogContent>
-            </AlertDialog>
-          </div>
-          <CardContent>
-            <Tabs defaultValue="Personas" className="w-full">
-              <TabsList>
-                <TabsTrigger value="Personas">Personas</TabsTrigger>
-                <TabsTrigger value="Equipos">Equipos</TabsTrigger>
-                <TabsTrigger value="Empresa">Empresa</TabsTrigger>
-              </TabsList>
-              <TabsContent value="Personas">
-                <Table>
-                  <TableHeader>
-                    <TableRow>
-                      <TableHead>Nombre del Documento</TableHead>
-                      <TableHead className="w-[100px] text-center" align="center">
-                        Multirecurso
-                      </TableHead>
-                      <TableHead className="w-[130px] text-center" align="center">
-                        Es especial?
-                      </TableHead>
-                      <TableHead className="w-[130px] text-center" align="center">
-                        Es mensual?
-                      </TableHead>
-                      <TableHead className="w-[100px] text-center" align="center">
-                        Vence
-                      </TableHead>
-                      <TableHead className="w-[100px] text-center" align="center">
-                        Mandatorio
-                      </TableHead>
-                      <TableHead className="w-[100px] text-center" align="center">
-                        Es privado?
-                      </TableHead>
-                      <TableHead className="w-[100px] text-center" align="center">
-                        Editar
-                      </TableHead>
-                    </TableRow>
-                  </TableHeader>
-                  <TableBody>
-                    {doc_personas
-                      ?.sort((a, b) => {
-                        if (a.company_id === null && b.company_id !== null) {
-                          return -1;
-                        } else if (a.company_id !== null && b.company_id === null) {
-                          return 1;
-                        } else {
-                          return a.name.localeCompare(b.name);
-                        }
-                      })
-                      ?.map((doc) => (
-                        <TableRow key={doc.id}>
-                          <TableCell className="font-medium">{doc.name}</TableCell>
-                          <TableCell align="center">{doc.multiresource ? 'Si' : 'No'}</TableCell>
-                          <TableCell align="center">{doc.special ? 'Si' : 'No'}</TableCell>
-                          <TableCell align="center">{doc.is_it_montlhy ? 'Si' : 'No'}</TableCell>
-                          <TableCell align="center">{doc.explired ? 'Si' : 'No'}</TableCell>
-                          <TableCell align="center">{doc.mandatory ? 'Si' : 'No'}</TableCell>
-                          <TableCell align="center">{doc.private ? 'Si' : 'No'}</TableCell>
-                          {doc.company_id && (
-                            <TableCell align="center">
-                              <EditModal Equipo={doc} />
-                            </TableCell>
-                          )}
-                        </TableRow>
-                      ))}
-                  </TableBody>
-                </Table>
-              </TabsContent>
-              <TabsContent value="Equipos">
-                <Table>
-                  <TableHeader>
-                    <TableRow>
-                      <TableHead>Nombre del Documento</TableHead>
-                      <TableHead className="w-[100px] text-center" align="center">
-                        Multirecurso
-                      </TableHead>
-                      <TableHead className="w-[130px] text-center" align="center">
-                        Es especial?
-                      </TableHead>
-                      <TableHead className="w-[130px] text-center" align="center">
-                        Es mensual?
-                      </TableHead>
-                      <TableHead className="w-[100px] text-center" align="center">
-                        Vence
-                      </TableHead>
-                      <TableHead className="w-[100px] text-center" align="center">
-                        Mandatorio
-                      </TableHead>
-                      <TableHead className="w-[100px] text-center" align="center">
-                        Es privado?
-                      </TableHead>
-                      <TableHead className="w-[100px] text-center" align="center">
-                        Editar
-                      </TableHead>
-                    </TableRow>
-                  </TableHeader>
-                  <TableBody>
-                    {doc_equipos
-                      ?.sort((a, b) => {
-                        if (a.company_id === null && b.company_id !== null) {
-                          return -1;
-                        } else if (a.company_id !== null && b.company_id === null) {
-                          return 1;
-                        } else {
-                          return a.name.localeCompare(b.name);
-                        }
-                      })
-                      ?.map((doc) => (
-                        <TableRow key={doc.id}>
-                          <TableCell className="font-medium">{doc.name}</TableCell>
-                          <TableCell align="center">{doc.multiresource ? 'Si' : 'No'}</TableCell>
-                          <TableCell align="center">{doc.special ? 'Si' : 'No'}</TableCell>
-                          <TableCell align="center">{doc.is_it_montlhy ? 'Si' : 'No'}</TableCell>
-                          <TableCell align="center">{doc.explired ? 'Si' : 'No'}</TableCell>
-                          <TableCell align="center">{doc.mandatory ? 'Si' : 'No'}</TableCell>
-                          <TableCell align="center">{doc.private ? 'Si' : 'No'}</TableCell>
-                          {doc.company_id && (
-                            <TableCell align="center">
-                              <EditModal Equipo={doc} />
-                            </TableCell>
-                          )}
-                        </TableRow>
-                      ))}
-                  </TableBody>
-                </Table>
-              </TabsContent>
-              <TabsContent value="Empresa">
-                <Table>
-                  <TableHeader>
-                    <TableRow>
-                      <TableHead>Nombre del Documento</TableHead>
-                      <TableHead className="w-[100px] text-center">Vence</TableHead>
-                      <TableHead className="w-[130px] text-center">Es mensual?</TableHead>
-                      <TableHead className="w-[100px] text-center">Es privado?</TableHead>
-                      <TableHead className="w-[100px] text-center">Mandatorio</TableHead>
-                      <TableHead className="w-[100px] text-center">Editar</TableHead>
-                    </TableRow>
-                  </TableHeader>
-                  <TableBody>
-                    {doc_empresa?.map((doc) => (
-                      <TableRow key={doc.id}>
-                        <TableCell className="font-medium">{doc.name}</TableCell>
-                        <TableCell align="center">{doc.explired ? 'Si' : 'No'}</TableCell>
-                        <TableCell align="center">{doc.is_it_montlhy ? 'Si' : 'No'}</TableCell>
-                        <TableCell align="center">{doc.private ? 'Si' : 'No'}</TableCell>
-                        <TableCell align="center">{doc.mandatory ? 'Si' : 'No'}</TableCell>
-                        {doc.company_id && (
-                          <TableCell align="center">
-                            <EditModal Equipo={doc} />
-                          </TableCell>
-                        )}
-                      </TableRow>
-                    ))}
-                  </TableBody>
-                </Table>
-              </TabsContent>
-            </Tabs>
-          </CardContent>
-        </Card>
+        <TypesDocumentsView />
       </TabsContent>
     </Tabs>
   );
